починила регистрацию, сломала получение токена

+++ backend/foodgram/users/views.py
@@ -1,4 +1,5 @@
 from djoser.views import UserViewSet as DjoserUserViewSet
+from rest_framework.authentication import TokenAuthentication
 from rest_framework import viewsets
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.filters import SearchFilter
@@ -8,9 +9,9 @@ from django.shortcuts import get_object_or_404
 
 from django.contrib.auth.models import User
 from rest_framework.decorators import action
-
-from .models import User, Follow
-from .serializers import FollowSerializer, CustomUserCreateSerializer, UserRegistrationSerializer
+from rest_framework.permissions import IsAuthenticatedOrReadOnly
+from .models import CustomUser, Follow
+from .serializers import FollowSerializer, CustomUserCreateSerializer
 from django.contrib.auth import authenticate, login, logout
 from rest_framework.authtoken.models import Token
 
@@ -18,13 +19,14 @@ POST_FILTER = 6
 
 
 class CustomUserViewSet(DjoserUserViewSet):
-    queryset = User.objects.all()
+    queryset = CustomUser.objects.all()
     serializer_class = CustomUserCreateSerializer
-    permission_classes = (IsAuthenticated,)
     pagination_class = LimitOffsetPagination
+    permission_classes = [IsAuthenticatedOrReadOnly]
     page_size = POST_FILTER
 
-    def list(self, request, *args, **kwargs):  # список пользователей
+    @action(detail=False, methods=['GET'])
+    def user_list(self, request, *args, **kwargs):  # список пользователей
         queryset = self.filter_queryset(self.get_queryset())
         total_count = queryset.count()
         page = self.paginate_queryset(queryset)
@@ -40,7 +42,7 @@ class CustomUserViewSet(DjoserUserViewSet):
         serializer = self.get_serializer(queryset, many=True)
         return Response(serializer.data)
 
-    @action(detail=False, methods=['get'])  # информация о текущем пользователе
+    @action(detail=False, methods=['POST'], permission_classes=(TokenAuthentication,))  # информация о текущем пользователе
     def current_user(self, request):
         user = request.user
         if user.is_authenticated:
@@ -49,7 +51,7 @@ class CustomUserViewSet(DjoserUserViewSet):
         else:
             return Response({'detail': 'Пользователь не авторизован.'})
 
-    @action(detail=False, methods=['get'], permission_classes=(AllowAny,))  # показывает профиль пользователя
+    @action(detail=False, methods=['GET'])  # показывает профиль пользователя
     def profile(self, request):
         user = request.user
         if not user.is_authenticared:
@@ -60,16 +62,15 @@ class CustomUserViewSet(DjoserUserViewSet):
         serializer = self.get_serializer(user_profile)
         return Response(serializer.data)
 
-    @action(detail=False, methods=['post'])  # регистрация пользователя
+    @action(detail=False, methods=['POST'])  # регистрация пользователя
     def register(self, request):
-        serializer = UserRegistrationSerializer(data=request.data)
+        serializer = CustomUserCreateSerializer(data=request.data)
         if serializer.is_valid():
             user = serializer.save()
             return Response(CustomUserCreateSerializer(user).data)
-
         return Response(serializer.errors)
 
-    @action(detail=False, methods=['post'], permission_classes=(IsAuthenticated,))  # смена пароля
+    @action(detail=False, methods=['post'])  # смена пароля
     def change_password(self, request):
         user = request.user
         current_password = request.data.get('current_password')
@@ -85,7 +86,7 @@ class CustomUserViewSet(DjoserUserViewSet):
         user.save()
         return Response()
 
-    @action(detail=False, methods=['post'], permission_classes=(AllowAny,))  # получение токена
+    @action(detail=False, methods=['post'])  # получение токена
     def obtain_auth_token(self, request):
         email = request.data.get('email')
         password = request.data.get('password')
@@ -96,13 +97,15 @@ class CustomUserViewSet(DjoserUserViewSet):
         user = authenticate(request, username=email, password=password)
 
         if user:
-            token, _ = Token.objects.get_or_create(user=user)
+            token, created  = Token.objects.get_or_create(user=user)
             login(request, user)
             return Response({'auth_token': token.key})
 
-    @action(detail=False, methods=['post'], permission_classes=(IsAuthenticated,))  # удаление токена
+    @action(detail=False, methods=['post'])  # удаление токена
     def delete_token(self, request):
         user = request.user
+        if not request.user.is_authenticated:
+            return Response({'detail': 'Пользователь не авторизован.'})
         try:
             token = Token.objects.get(user=user)
             token.delete()
@@ -114,6 +117,7 @@ class CustomUserViewSet(DjoserUserViewSet):
 
 class FollowViewSet(viewsets.ModelViewSet):
     """ViewSet для подписки авторизованного пользователя."""
+    queryset = CustomUser.objects.none()
     serializer_class = FollowSerializer
     pagination_class = LimitOffsetPagination
     page_size = POST_FILTER
@@ -128,7 +132,7 @@ class FollowViewSet(viewsets.ModelViewSet):
 
     @action(detail=False, methods=['post'])
     def create_subscription(self, request, pk=None):  # создание подписки
-        user_to_sub_to = get_object_or_404(User, pk=pk)
+        user_to_sub_to = get_object_or_404(CustomUser, pk=pk)
         user = request.user
         serializer = self.get_serializer(data={'user': user.id, 'following': user_to_sub_to.id})
         serializer.is_valid(raise_exception=True)
@@ -138,7 +142,7 @@ class FollowViewSet(viewsets.ModelViewSet):
     def delete_subscription(self, request, pk=None):  # Отписка
         if not request.user.is_authenticated:
             return Response({'detail': 'Пользователь не авторизован.'})
-        user_to_unsub = get_object_or_404(User, pk=pk)
+        user_to_unsub = get_object_or_404(CustomUser, pk=pk)
         user = request.user
 
         try:
